+++ internal/git/command_description.go
@@ -53,12 +53,12 @@ var commandDescriptions = map[string]commandDescription{
 		flags: scNoEndOfOptions,
 	},
 	"clone": {
-		opts: append([]GlobalOption{
+		opts: append(append([]GlobalOption{
 			// See "init" for why we set the template directory to the empty string.
 			ConfigPair{Key: "init.templateDir", Value: ""},
 			// See "fetch" for why we disable following redirects.
 			ConfigPair{Key: "http.followRedirects", Value: "false"},
-		}, packConfiguration()...),
+		}, packConfiguration()...), fsckConfiguration("fetch")...),
 	},
 	"commit": {
 		flags: 0,
@@ -298,12 +298,12 @@ var commandDescriptions = map[string]commandDescription{
 	},
 	"upload-pack": {
 		flags: scNoRefUpdates,
-		opts: append(append([]GlobalOption{
+		opts: append([]GlobalOption{
 			ConfigPair{Key: "uploadpack.allowFilter", Value: "true"},
 			// Enables the capability to request individual SHA1's from the
 			// remote repo.
 			ConfigPair{Key: "uploadpack.allowAnySHA1InWant", Value: "true"},
-		}, hiddenUploadPackRefPrefixes()...), packConfiguration()...),
+		}, packConfiguration()...),
 	},
 	"version": {
 		flags: scNoRefUpdates,