+++ internal/gitaly/service/repository/archive_test.go
@@ -5,6 +5,7 @@ package repository
 import (
 	"archive/zip"
 	"bytes"
+	"context"
 	"fmt"
 	"io"
 	"os"
@@ -19,6 +20,7 @@ import (
 	"gitlab.com/gitlab-org/gitaly/v15/internal/gitaly/config"
 	"gitlab.com/gitlab-org/gitaly/v15/internal/gitlab"
 	"gitlab.com/gitlab-org/gitaly/v15/internal/helper/text"
+	"gitlab.com/gitlab-org/gitaly/v15/internal/metadata/featureflag"
 	"gitlab.com/gitlab-org/gitaly/v15/internal/testhelper"
 	"gitlab.com/gitlab-org/gitaly/v15/internal/testhelper/testcfg"
 	"gitlab.com/gitlab-org/gitaly/v15/internal/testhelper/testserver"
@@ -165,10 +167,13 @@ func TestGetArchive_success(t *testing.T) {
 	}
 }
 
-func TestGetArchive_includeLfsBlobs(t *testing.T) {
+func TestGetArchiveIncludeLfsBlobs(t *testing.T) {
 	t.Parallel()
+	testhelper.NewFeatureSets(featureflag.UploadPackHideRefs).Run(t, testGetArchiveIncludeLfsBlobs)
+}
 
-	ctx := testhelper.Context(t)
+func testGetArchiveIncludeLfsBlobs(t *testing.T, ctx context.Context) {
+	t.Parallel()
 
 	defaultOptions := gitlab.TestServerOptions{
 		SecretToken: secretToken,